% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_is_duplicated}
\alias{add_is_duplicated}
\title{Check if rowname is duplicated and add binary is_duplicated
Helper function}
\usage{
add_is_duplicated(dat, rowname)
}
\arguments{
\item{dat}{data frame}

\item{rowname}{rowname to check as character}
}
\description{
Check if rowname is duplicated and add binary is_duplicated
Helper function
}
