% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{attach_gene_symbol_from_entrez}
\alias{attach_gene_symbol_from_entrez}
\title{Add Gene Symbol from EntrezID}
\usage{
attach_gene_symbol_from_entrez(dat)
}
\arguments{
\item{dat}{data frame with EnzrezID}
}
\description{
Add Gene Symbol from EntrezID
}
