% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{attach_goterm_genecount}
\alias{attach_goterm_genecount}
\title{Add count for each GO-term and calculate overlap ratio of significant genes}
\usage{
attach_goterm_genecount(dat)
}
\arguments{
\item{dat}{clusterProfiler GO-term result DOSE}
}
\description{
Add count for each GO-term and calculate overlap ratio of significant genes
}
