% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{bind_goterm_table}
\alias{bind_goterm_table}
\title{Combine goterms into a single data frame}
\usage{
bind_goterm_table(dat)
}
\arguments{
\item{dat}{clusterProfiler result}
}
\value{
data frame with goterms and GO-term source column
}
\description{
Combine goterms into a single data frame
}
