% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{emap_plot_facet_category}
\alias{emap_plot_facet_category}
\title{Combine emap plots into one ggplot2 plot}
\usage{
emap_plot_facet_category(dat, n = 50)
}
\arguments{
\item{dat}{clusterProfiler result}

\item{n}{See enrichplot::emapplot}
}
\value{
ggplot2 plot
}
\description{
Combine emap plots into one ggplot2 plot
}
