% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ensembl_to_entrez}
\alias{ensembl_to_entrez}
\title{Gets Entrez IDs and add it to dat}
\usage{
ensembl_to_entrez(dat, ensembl_id_name,
  keep_only_rows_with_entrez = TRUE, drop_duplicates = TRUE)
}
\arguments{
\item{dat}{Data frame with ensembl identifier}

\item{ensembl_id_name}{Name of column containing the ensembl identifier}

\item{keep_only_rows_with_entrez}{Only keep rows for which entrez IDs could be found}

\item{drop_duplicates}{Often there is a n:1 relationship between Entrez-IDs and Ensembl-IDs. If this value is true, only keep the first hit}
}
\description{
Gets Entrez IDs and add it to dat
}
