% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{export_go_terms_to_excel}
\alias{export_go_terms_to_excel}
\title{Export list of GO terms to Excel file
TODO: Redo this!}
\usage{
export_go_terms_to_excel(ontologies, path)
}
\arguments{
\item{ontologies}{List of clusterProfiler enrichResult objects with elements BP, MF and CC}

\item{path}{Path to exported excel file}
}
\description{
Export list of GO terms to Excel file
TODO: Redo this!
}
