% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_go_all_ontologies}
\alias{get_go_all_ontologies}
\title{Get GO-terms for all ontologies}
\usage{
get_go_all_ontologies(dat, significance_cutoff = 0.05)
}
\arguments{
\item{dat}{Data frame containing columns \code{pValue} and \code{EntrezID}}

\item{significance_cutoff}{Cutoff to consider genes as significant}
}
\description{
Get GO-terms for all ontologies
}
