% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_go_all_ontologies_2}
\alias{get_go_all_ontologies_2}
\title{Get GO-terms for all ontologies. This method requires a \code{significant} flag in the data frame}
\usage{
get_go_all_ontologies_2(dat, use_background = TRUE, species = "MUS")
}
\arguments{
\item{dat}{Data frame containing columns \code{significant} and \code{EntrezID}}

\item{use_background}{Use gene list as background instead of using all genes as background}

\item{species}{Either "HUM" or "MUS"}
}
\description{
Get GO-terms for all ontologies. This method requires a \code{significant} flag in the data frame
}
