% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_go_all_ontologies_helper}
\alias{get_go_all_ontologies_helper}
\title{Helper function getting all GO-terms based on vectors of Entrez-IDs}
\usage{
get_go_all_ontologies_helper(
  significant_entrezgenes,
  background_entrezgenes,
  use_background
)
}
\arguments{
\item{significant_entrezgenes}{Vector of significant genes as Entrez-IDs}

\item{background_entrezgenes}{Vector of background genes as Entrez-IDs}

\item{use_background}{Use background entrezgenes Vector}
}
\description{
Helper function getting all GO-terms based on vectors of Entrez-IDs
}
