% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_kegg_table}
\alias{get_kegg_table}
\title{Get Table containing Kegg pathway information \code{pathway}, \code{ensembl_gene_id}, \code{kegg_name}}
\usage{
get_kegg_table(species = "MUS")
}
\arguments{
\item{species}{Species string, either "MUS" or "HUM"}
}
\value{
Table containing Kegg pathway information \code{pathway}, \code{ensembl_gene_id}, \code{kegg_name}
}
\description{
Get Table containing Kegg pathway information \code{pathway}, \code{ensembl_gene_id}, \code{kegg_name}
}
\examples{
get_kegg_table("MUS")
}
