% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_named_fc_vector}
\alias{get_named_fc_vector}
\title{Get named fc vector. Prepare input data as required by GSE}
\usage{
get_named_fc_vector(dat)
}
\arguments{
\item{dat}{Dataframe with columns \code{EntrezID} and \code{fc}}
}
\description{
Get named fc vector. Prepare input data as required by GSE
}
