% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{overlap_percentage_plot}
\alias{overlap_percentage_plot}
\title{Plot top n GO-terms w.r.t percentage of significant genes
Plot adapted from https://yulab-smu.github.io/clusterProfiler-book/chapter13.html}
\usage{
overlap_percentage_plot(dat, n = 20)
}
\arguments{
\item{dat}{clusterProfiler GO-term table with attached GOTermGeneCount and Percent_Significant column}

\item{n}{Number of GO-terms to label}
}
\description{
Plot top n GO-terms w.r.t percentage of significant genes
Plot adapted from https://yulab-smu.github.io/clusterProfiler-book/chapter13.html
}
\examples{
  dat_goterms$Biological_Process \%>\%
    mygo::attach_goterm_genecount() \%>\%
    mygo::overlap_percentage_plot()
}
