% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{overlap_percentage_plot_facet_category}
\alias{overlap_percentage_plot_facet_category}
\title{Wrapper for mygo::overlap_percentage_plot to facet all GO categories}
\usage{
overlap_percentage_plot_facet_category(dat, order_by = "significance", n = 20)
}
\arguments{
\item{dat}{clusterProfiler result}

\item{order_by}{See mygo::overlap_percentage_plot}

\item{n}{See mygo::overlap_percentage_plot}
}
\value{
ggplot2 object
}
\description{
Wrapper for mygo::overlap_percentage_plot to facet all GO categories
}
