% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{overlap_scatterplot}
\alias{overlap_scatterplot}
\title{Plot total gene count against percentage of significant genes}
\usage{
overlap_scatterplot(dat, n = 15)
}
\arguments{
\item{dat}{clusterProfiler GO-term table with attached GOTermGeneCount and Percent_Significant column}

\item{n}{Number of GO-terms to label}
}
\description{
Plot total gene count against percentage of significant genes
}
\examples{
  dat_goterms$Biological_Process \%>\%
    mygo::attach_goterm_genecount(species = "MUS") \%>\%
    overlap_scatterplot
}
