% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{perform_enrichGO}
\alias{perform_enrichGO}
\title{Perform a clusterPrfiler enrichGO analysis}
\usage{
perform_enrichGO(ontology, entrezgenes)
}
\arguments{
\item{ontoloty}{Can be either "BP", "CC", "MF"}

\item{enztezgenes}{List of entrezgenes to use for GO analysis}
}
\description{
Perform a clusterPrfiler enrichGO analysis
}
