% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_all_ontologies}
\alias{plot_all_ontologies}
\title{Print a couple of standard plots for provided GO ontologies}
\usage{
plot_all_ontologies(ontologies, fc_symbol)
}
\arguments{
\item{ontologies}{List of clusterProfiler enrichResult objects with elements BP, MF and CC}

\item{fc_symbol}{double vector containing the fold changes named by symbol}
}
\description{
Print a couple of standard plots for provided GO ontologies
}
