% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_go_terms_all_ontologies}
\alias{plot_go_terms_all_ontologies}
\title{Print a couple of standard plots for provided GO ontologies}
\usage{
plot_go_terms_all_ontologies(ontologies)
}
\arguments{
\item{ontologies}{List of clusterProfiler enrichResult objects with elements BP, MF and CC}
}
\description{
Print a couple of standard plots for provided GO ontologies
}
