% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_gos_normal}
\alias{plot_gos_normal}
\title{Print a series of clusterProfiler plots}
\usage{
plot_gos_normal(go_terms)
}
\arguments{
\item{go_terms}{clusterProfiler enrichResult object}
}
\description{
Print a series of clusterProfiler plots
}
