% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_kegg}
\alias{plot_kegg}
\title{Create KEGG pathway as PNG}
\usage{
plot_kegg(dat, pathway, species = "mmu")
}
\arguments{
\item{dat}{dataframe containing columns EntrezID and log2FoldChange}

\item{pathway}{kegg pathway name}

\item{species}{kegg species ID}
}
\value{
png object
}
\description{
Create KEGG pathway as PNG
}
\examples{
  mygo::plot_kegg(kegg_dat, pathway = "mmu04110") \%>\% mygo::plot_png()
}
