% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_kegg_pathway}
\alias{plot_kegg_pathway}
\title{Create KEGG pathway as PNG}
\usage{
plot_kegg_pathway(dat, pathway, species = "mmu")
}
\arguments{
\item{dat}{dataframe containing columns EntrezID and log2FoldChange}

\item{pathway}{kegg pathway name}

\item{species}{kegg species ID}
}
\description{
Create KEGG pathway as PNG
}
