% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_terms_gse}
\alias{plot_terms_gse}
\title{Plot GSE terms}
\usage{
plot_terms_gse(gse_terms, fc_symbol)
}
\arguments{
\item{gse_terms}{clusterProfiler gse terms object}

\item{fc_symbol}{double vector containing the fold changes named by symbol}
}
\description{
Plot GSE terms
}
