% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_terms_normal}
\alias{plot_terms_normal}
\title{Print a series of clusterProfiler plots}
\usage{
plot_terms_normal(go_terms, fc_symbol)
}
\arguments{
\item{go_terms}{clusterProfiler enrichResult object}

\item{fc_symbol}{double vector containing the fold changes named by symbol}
}
\description{
Print a series of clusterProfiler plots
}
