% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print_goterm_as_datatable}
\alias{print_goterm_as_datatable}
\title{Helper function, takes clusterProfiler GO-term result and prints it as DT widget}
\usage{
print_goterm_as_datatable(dat)
}
\arguments{
\item{dat}{clusterProfiler GO-term result}
}
\description{
Helper function, takes clusterProfiler GO-term result and prints it as DT widget
}
