% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_goterms}
\alias{run_goterms}
\title{Run GO-Term analysis and retrieve plots
This is for usage in custom analysis scripts}
\usage{
run_goterms(
  dat,
  significance_cutoff = 0.05,
  fc_cutoff = 0,
  use_background = FALSE,
  simplify = TRUE,
  simplify_cutoff = 0.7,
  run_kegg = FALSE,
  species = "MUS"
)
}
\arguments{
\item{dat}{Table containing \code{ensembl_gene_id}, \code{padj} and optionally \code{log2FoldChange}}

\item{significance_cutoff}{Adjusted p-value cutoff}

\item{fc_cutoff}{Fold change cutoff}

\item{use_background}{Use whole table as gene background}

\item{simplify}{Summarise GO-terms}

\item{simplify_cutoff}{When summarising, use this cutoff}

\item{run_kegg}{Run KEGG pathway analysis. Adds columns \code{kegg} and \code{kegg_dt}}

\item{species}{Species string, either "MUS" or "HUM"}
}
\value{
list containing plots and table outputs
}
\description{
Run GO-Term analysis and retrieve plots
This is for usage in custom analysis scripts
}
