% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{simplify_ontologies}
\alias{simplify_ontologies}
\title{Simplify GO terms}
\usage{
simplify_ontologies(ontologies)
}
\arguments{
\item{ontologies}{List of clusterProfiler enrichResult objects with elements BP, MF and CC}
}
\description{
Simplify GO terms
}
