% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Return a volcano plot with annotated top_n values}
\usage{
volcano_plot(dat, label_top_n = 20)
}
\arguments{
\item{go_terms}{clusterProfiler enrichResult object}
}
\description{
Return a volcano plot with annotated top_n values
}
