% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{createHTMLReport}
\alias{createHTMLReport}
\title{Make GO-Term analysis and print out Report}
\usage{
createHTMLReport(dat, output_path, save_excel = TRUE,
  significance_cutoff = 0.05, dev = FALSE)
}
\arguments{
\item{dat}{Dataframe containing variables \code{q_value} and \code{EntrezID} or \code{EnsemblID}}

\item{output_path}{Path to HTML output}

\item{save_excel}{Save GO-terms to Excel files}

\item{significance_cutoff}{Specify the cutoff which entries are considered significant}
}
\value{
cummeRbund cuff object
}
\description{
Make GO-Term analysis and print out Report
}
