% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{createHTMLReport}
\alias{createHTMLReport}
\title{Make GO-Term analysis and print out Report}
\usage{
createHTMLReport(dat, output_path, save_excel = TRUE,
  significance_cutoff = 0.05, do_gse = TRUE,
  simplify_ontologies = TRUE, use_background = TRUE,
  store_r_objects = TRUE, simplify_cutoff = 0.7,
  save_plots_as_pdf = TRUE)
}
\arguments{
\item{dat}{Dataframe containing variables \code{q_value} and \code{EntrezID} or \code{EnsemblID}}

\item{output_path}{Path to HTML output}

\item{save_excel}{Save GO-terms to Excel files}

\item{significance_cutoff}{Specify the cutoff which entries are considered significant}

\item{do_gse}{Conduct GSE analysis}

\item{simplify_ontologies}{Run time-consuming ontology simplification}

\item{use_background}{Use background genes for the analysis}

\item{store_r_objects}{Store clusterProfiler R objects as compressed RDS files to output_path}

\item{simplify_cutoff}{See clusterProfiler::simplify}

\item{save_plots_as_pdf}{Export all plots as PDF to output_path/plots}
}
\value{
cummeRbund cuff object
}
\description{
Make GO-Term analysis and print out Report
}
