% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{emap_plot}
\alias{emap_plot}
\title{Create clusterProfiler EMA plot from go term}
\usage{
emap_plot(go_terms, title)
}
\arguments{
\item{go_terms}{clusterProfiler enrichResult object}

\item{title}{ggtitle of the plot}
}
\description{
Create clusterProfiler EMA plot from go term
}
