% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{export_go_terms_to_excel}
\alias{export_go_terms_to_excel}
\title{Export list of GO terms to Excel file}
\usage{
export_go_terms_to_excel(
  go_ontologies,
  go_ontologies_simple,
  gse_ontologies,
  kegg_ontologies,
  path,
  species = "MUS"
)
}
\arguments{
\item{go_ontologies}{List of clusterProfiler enrichResult objects with elements BP, MF and CC}

\item{go_ontologies_simple}{List of clusterProfiler enrichResult objects with elements BP, MF and CC after being simplified}

\item{gse_ontologies}{GSE enrichResult}

\item{kegg_ontologies}{KEGG enrichResult}

\item{path}{Path to exported excel file}

\item{species}{Either "MUS" or "HUM"}
}
\description{
Export list of GO terms to Excel file
}
