% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_go_all_ontologies}
\alias{get_go_all_ontologies}
\title{Get GO-terms for all ontologies}
\usage{
get_go_all_ontologies(
  dat,
  significance_cutoff = 0.05,
  fc_cutoff = 0,
  use_background = TRUE,
  species = "MUS"
)
}
\arguments{
\item{dat}{Data frame containing columns \code{q_value}, \code{EntrezID} and optionally \code{fc}}

\item{significance_cutoff}{Cutoff to consider genes as significant}

\item{fc_cutoff}{Fold change cutoff to consider genes as significant}

\item{use_background}{Use gene list as background instead of using all genes as background}

\item{species}{Either "HUM" or "MUS"}
}
\description{
Get GO-terms for all ontologies
}
