% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_gse_all_ontologies}
\alias{get_gse_all_ontologies}
\title{Get GSE-terms for all ontologies}
\usage{
get_gse_all_ontologies(dat)
}
\arguments{
\item{dat}{Data frame containing columns \code{pValue} and \code{EntrezID}}
}
\description{
Get GSE-terms for all ontologies
}
