% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_gse_all_ontologies}
\alias{get_gse_all_ontologies}
\title{Get GSE-terms for all ontologies}
\usage{
get_gse_all_ontologies(
  dat,
  p_cutoff = 0.05,
  set_readable = TRUE,
  simplify = FALSE,
  simplify_cutoff = 0.7
)
}
\arguments{
\item{dat}{Table containing columns \code{fc} and \code{EntrezID}}

\item{p_cutoff}{P-value cutoff for GSEA}

\item{set_readable}{Replace EntrezIDs by readable gene names. This causes problems with ridgeplot}

\item{simplify}{Simplify terms}

\item{simplify_cutoff}{cutoff value of clusterProfiler::simplify()}
}
\description{
Get GSE-terms for all ontologies
}
