% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{perform_gseGO}
\alias{perform_gseGO}
\title{Perform a clusterProfiler gseGO analysis}
\usage{
perform_gseGO(ontology, fc, p_cutoff = 0.05, set_readable = TRUE)
}
\arguments{
\item{ontology}{Can be either "BP", "CC", "MF"}

\item{fc}{Named vector of foldchanges (name denotes Entrez ID)}

\item{p_cutoff}{P-value cutoff for GSEA}

\item{set_readable}{Replace EntrezIDs by readable gene names. This causes problems with ridgeplot}
}
\description{
Perform a clusterProfiler gseGO analysis
}
