% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{perform_gseKEGG}
\alias{perform_gseKEGG}
\title{Perform a clusterProfiler gseKEGG analysis}
\usage{
perform_gseKEGG(ontology, fc)
}
\arguments{
\item{fc}{Named vector of foldchanges (name denotes Entrez ID)}

\item{ontoloty}{Can be either "BP", "CC", "MF"}
}
\description{
Perform a clusterProfiler gseKEGG analysis
}
