% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_goterms}
\alias{run_goterms}
\title{Run GO-Term analysis and retrieve plots
This is for usage in custom analysis scripts}
\usage{
run_goterms(
  dat,
  significance_cutoff = 0.05,
  use_background = FALSE,
  simplify = TRUE,
  simplify_cutoff = 0.7
)
}
\arguments{
\item{dat}{Table containing \code{ensembl_gene_id} and \code{padj}}

\item{significance_cutoff}{Adjusted p-value cutoff}

\item{use_background}{Use whole table as gene background}

\item{simplify}{Summarise GO-terms}

\item{simplify_cutoff}{When summarising, use this cutoff}
}
\value{
list containing plots and table outputs
}
\description{
Run GO-Term analysis and retrieve plots
This is for usage in custom analysis scripts
}
